+++ ChangeLog.md
@@ -1,19 +1,33 @@
+# Version 0.2.0 
 
+#### Change models
++ root.py : Add 1 more helper functions
++ Swarm_based: Added
+    * PSO
+    * BFO, ABFOLS (Adaptive version of BFO)
+    * CSO
+    * ABC
+    * WOA
+
+#### Change others
++ models_history.csv: Adding history of meta-heuristic algorithms
+    
+---------------------------------------------------------------------
 # Version 0.1.1 
 
-### Change models
+#### Change models
 + root.py : Add more helper functions
 + Evolutionary_based
     * GA : Change the format of input parameters
     * DE : Change the format of input parameters
-### Change others
+#### Change others
 + Examples: Adding more complex examples
 + Library: "Opfunu" update the latest verion 0.4.3
     
 ---------------------------------------------------------------------
 # Version 0.1.0 (First version)
 
-### Changed models
+#### Changed models
 + root.py (Very first file, the root of all algorithms)
 + Evolutionary_based
     * GA